fix(agents): move eval test-support modules under testing/
The all_blueprints registry scan excludes /testing/ paths (the same
convention as dimos/agents/testing/); the eval's observation-only
modules are test scaffolding and do not belong in the module registry.
